+++ guillotina/tests/test_postgres.py
@@ -2,11 +2,13 @@ from guillotina.api.container import create_container
 from guillotina.component import get_adapter
 from guillotina.content import Folder
 from guillotina.db.interfaces import IVacuumProvider
+from guillotina.db.locks import lock_object_for_write
 from guillotina.db.storages.cockroach import CockroachStorage
 from guillotina.db.storages.pg import PostgresqlStorage
 from guillotina.db.transaction_manager import TransactionManager
 from guillotina.exceptions import ConflictError
 from guillotina.exceptions import ConflictIdOnContainer
+from guillotina.exceptions import ObjectLockedError
 from guillotina.tests import mocks
 from guillotina.tests.utils import create_content
 from unittest.mock import Mock
@@ -44,11 +46,15 @@ async def cleanup(aps):
 
 
 async def get_aps(postgres, pool_size=16, autovacuum=True):
-    dsn = "postgres://postgres:postgres@{}:{}/guillotina".format(postgres[0], postgres[1])
+    dsn = "postgres://postgres:postgres@{}:{}/guillotina".format(
+        postgres[0], postgres[1]
+    )
     klass = PostgresqlStorage
     if DATABASE == "cockroachdb":
         klass = CockroachStorage
-        dsn = "postgres://root:@{}:{}/guillotina?sslmode=disable".format(postgres[0], postgres[1])
+        dsn = "postgres://root:@{}:{}/guillotina?sslmode=disable".format(
+            postgres[0], postgres[1]
+        )
     aps = klass(
         dsn=dsn,
         name="db",
@@ -133,6 +139,37 @@ async def test_restart_connection_pg(db, dummy_guillotina):
 
         await aps.remove()
         await cleanup(aps)
+
+
+@pytest.mark.skipif(DATABASE == "DUMMY", reason="Not for dummy db")
+async def test_object_lock_for_write(db, dummy_guillotina):
+    aps = await get_aps(db)
+    tm = TransactionManager(aps)
+
+    async with tm:
+        txn1 = await tm.begin()
+        ob = create_content()
+        txn1.register(ob)
+        await tm.commit(txn=txn1)
+
+        txn1 = await tm.begin()
+        async with lock_object_for_write(ob.__uuid__, retries=1, delay=0):
+            txn2 = await tm.begin()
+            try:
+                with pytest.raises(ObjectLockedError):
+                    async with lock_object_for_write(ob.__uuid__, retries=2, delay=0):
+                        pass
+            finally:
+                await tm.abort(txn=txn2)
+        await tm.abort(txn=txn1)
+
+        txn3 = await tm.begin()
+        async with lock_object_for_write(ob.__uuid__, retries=1, delay=0):
+            pass
+        await tm.abort(txn=txn3)
+
+    await aps.remove()
+    await cleanup(aps)
 
 
 @pytest.mark.skipif(
@@ -234,8 +271,12 @@ async def test_delete_resource_deletes_blob(db, dummy_guillotina):
         await cleanup(aps)
 
 
-@pytest.mark.skipif(DATABASE in ("cockroachdb", "DUMMY"), reason="Cockroach not support resolve...")
-async def test_should_raise_conflict_error_when_editing_diff_data_with_resolve_strat(db, dummy_guillotina):
+@pytest.mark.skipif(
+    DATABASE in ("cockroachdb", "DUMMY"), reason="Cockroach not support resolve..."
+)
+async def test_should_raise_conflict_error_when_editing_diff_data_with_resolve_strat(
+    db, dummy_guillotina
+):
     aps = await get_aps(db)
     with TransactionManager(aps) as tm, await tm.begin() as txn:
         ob = create_content()
@@ -266,7 +307,9 @@ async def test_should_raise_conflict_error_when_editing_diff_data_with_resolve_s
         await cleanup(aps)
 
 
-@pytest.mark.skipif(DATABASE in ("cockroachdb", "DUMMY"), reason="Cockroach not support resolve...")
+@pytest.mark.skipif(
+    DATABASE in ("cockroachdb", "DUMMY"), reason="Cockroach not support resolve..."
+)
 async def test_should_resolve_conflict_error(db, dummy_guillotina):
     aps = await get_aps(db)
     with TransactionManager(aps) as tm, await tm.begin() as txn:
@@ -297,7 +340,9 @@ async def test_should_resolve_conflict_error(db, dummy_guillotina):
 
 
 @pytest.mark.skipif(DATABASE in ("DUMMY",), reason="DUMMY not support simple...")
-async def test_should_raise_conflict_error_on_concurrent_insert(db, container_requester):
+async def test_should_raise_conflict_error_on_concurrent_insert(
+    db, container_requester
+):
     async with container_requester as requester:
         with requester.db.get_transaction_manager() as tm:
             txn = await tm.begin()
@@ -324,7 +369,9 @@ async def test_should_raise_conflict_error_on_concurrent_insert(db, container_re
 
 
 @pytest.mark.skipif(DATABASE in ("DUMMY",), reason="DUMMY not support simple...")
-async def test_should_raise_conflict_error_on_concurrent_update(db, container_requester):
+async def test_should_raise_conflict_error_on_concurrent_update(
+    db, container_requester
+):
     async with container_requester as requester:
         with requester.db.get_transaction_manager() as tm:
             txn = await tm.begin()
@@ -528,7 +575,9 @@ async def test_iterate_keys(db, dummy_guillotina):
         await tm.abort(txn=txn)
 
 
-@pytest.mark.skipif(DATABASE in ("cockroachdb", "DUMMY"), reason="Cockroach does not like this test...")
+@pytest.mark.skipif(
+    DATABASE in ("cockroachdb", "DUMMY"), reason="Cockroach does not like this test..."
+)
 async def test_handles_asyncpg_trying_savepoints(db, dummy_guillotina):
     aps = await get_aps(db)
     tm = TransactionManager(aps)
@@ -536,7 +585,9 @@ async def test_handles_asyncpg_trying_savepoints(db, dummy_guillotina):
     for conn in tm._storage.pool._queue._queue:
         if conn._con is None:
             await conn.connect()
-        conn._con._top_xact = asyncpg.transaction.Transaction(conn._con, "read_committed", False, False)
+        conn._con._top_xact = asyncpg.transaction.Transaction(
+            conn._con, "read_committed", False, False
+        )
 
     with await tm.begin() as txn, tm:
         # then, try doing stuff...
@@ -558,7 +609,9 @@ async def test_handles_asyncpg_trying_savepoints(db, dummy_guillotina):
         await cleanup(aps)
 
 
-@pytest.mark.skipif(DATABASE in ("cockroachdb", "DUMMY"), reason="Cockroach does not like this test...")
+@pytest.mark.skipif(
+    DATABASE in ("cockroachdb", "DUMMY"), reason="Cockroach does not like this test..."
+)
 async def test_handles_asyncpg_trying_txn_with_manual_txn(db, dummy_guillotina):
     aps = await get_aps(db)
     tm = TransactionManager(aps)